+++ src/quacc/__init__.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING
 from ase.atoms import Atoms
 
 from quacc.settings import QuaccSettings
+from quacc.utils.dicts import Remove
 from quacc.wflow_tools.customizers import redecorate, strip_decorator
 from quacc.wflow_tools.decorators import Flow, Job, Subflow, flow, job, subflow
 
@@ -22,6 +23,7 @@ __all__ = [
     "Subflow",
     "redecorate",
     "strip_decorator",
+    "Remove",
 ]